// hr-frontend/src/lib/api/generated/model/historyEntry.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface HistoryEntry {
  employmentId?: number;
  orgUnitName?: string;
  positionTitle?: string;
  reportingToName?: string;
  startDate?: string;
  endDate?: string;
  current?: boolean;
}